UserNavigation is now a component
