+++ src/components/common/UserNavigation.jsx
@@ -1,0 +1,33 @@
+import React, { useState } from 'react'
+/*    Components    */
+import SearchBar from '../common/SearchBar'
+import RoundButton from '../common/RoundButton'
+import NavListContainer from '../common/navBar/NavListContainer'
+import LogoScript from '../common/LogoScript'
+/*       Icons        */
+import HeartDark from '../../assets/common/HeartDark.svg'
+import Location from '../../assets/common/Location.svg'
+import CartDark from '../../assets/common/CartDark.svg'
+import Person from '../../assets/common/Person.svg'
+
+const UserNavigation = () => {
+  const [query, setQuery] = useState('')
+  const searchResult = (e) => {
+    setQuery(e.target.value)
+  }
+  return (
+    <>
+      <div className="flex pt-3 px-[112px]">
+        <LogoScript />
+        <SearchBar onGetData={searchResult} query={query} />
+        <RoundButton onGetLogo={HeartDark} onGetText="المفضلة" onGetPath="/favorites" />
+        <RoundButton onGetLogo={Location} onGetText="أقرب صيدلية" onGetPath="/location" />
+        <RoundButton onGetLogo={CartDark} onGetText="العربة" onGetPath="/cart" />
+        <RoundButton onGetLogo={Person} onGetText="حسابي" onGetPath="/profile" />
+      </div>
+      <NavListContainer />
+    </>
+  )
+}
+
+export default UserNavigation
