+++ src/components/pages/ProductPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+/*       API       */
+import * as ProductsAPI from '../../utils/productsAPI' //getProduct
+
+const ProductPage = () => {
+  return (
+    <div>
+        
+    </div>
+  )
+}
+
+export default ProductPage
