+++ src/components/pages/favourites/Favourites.jsx
@@ -1,26 +1,15 @@
 import React, { useEffect, useState } from 'react'
 /*    Components    */
-import SearchBar from '../../common/SearchBar'
-import RoundButton from '../../common/RoundButton'
-import NavListContainer from '../../common/navBar/NavListContainer'
+import UserNavigation from '../../common/UserNavigation'
 import PageEmpty from '../../common/PageEmpty'
 import FavouritesContent from './FavouritesContent'
-import LogoScript from '../../common/LogoScript'
 /*    Icons    */
-import HeartDark from '../../../assets/common/HeartDark.svg'
 import HeartLight from '../../../assets/common/HeartLight.svg'
-import Location from '../../../assets/common/Location.svg'
-import CartDark from '../../../assets/common/CartDark.svg'
-import Person from '../../../assets/common/Person.svg'
 /*     API     */
 import * as ProductsAPI from '../../../utils/productsAPI'
 import * as UsersAPI from '../../../utils/usersAPI'
 
 const Favourites = () => {
-  const [query, setQuery] = useState('')
-  const searchResult = (e) => {
-    setQuery(e.target.value)
-  }
   const [items, setItems] = useState([])
   useEffect(() => {
     const updateItems = async () => {
@@ -36,15 +25,7 @@ const Favourites = () => {
 
   return (
     <>
-      <div className="flex pt-3 px-[112px]">
-        <LogoScript />
-        <SearchBar onGetData={searchResult} query={query} />
-        <RoundButton onGetLogo={HeartDark} onGetText="المفضلة" onGetPath="/favorites" />
-        <RoundButton onGetLogo={Location} onGetText="أقرب صيدلية" onGetPath="/location" />
-        <RoundButton onGetLogo={CartDark} onGetText="العربة" onGetPath="/cart" />
-        <RoundButton onGetLogo={Person} onGetText="حسابي" onGetPath="/profile" />
-      </div>
-      <NavListContainer />
+      <UserNavigation />
       <div>
         {items.length ? (
           <FavouritesContent
